requirements.txt
# Reddit API
praw>=7.7.0

# AI APIs
openai>=1.0.0
anthropic>=0.3.0
boto3>=1.26.0  # For AWS Bedrock

# Web Framework
streamlit>=1.25.0

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0
markdown>=3.4.0

# Data handling
pandas>=2.0.0
pyyaml>=6.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Logging and monitoring
structlog>=23.1.0

# Caching
diskcache>=5.6.0

# AWS CDK (optional, for infrastructure)
# aws-cdk-lib>=2.90.0
# constructs>=10.0.0